+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 - added addional config for `.nav_stacked`
 - added color-modifier `.hover:is-ink-#`
 - added theme-modifier `.hover:is-theme-#`
+- added theme-modifier `.hover:is-elevated-#`
 - added `data-col="is-fit"`
 - added grid config `$grid-small-gutter`
 - added `is-round` modifier